+++ src/main/java/io/github/vampirestudios/artifice/api/ArtificeAssetPackEntrypoint.java
@@ -1,0 +1,8 @@
+package io.github.vampirestudios.artifice.api;
+
+import net.minecraft.resources.ResourceLocation;
+
+public interface ArtificeAssetPackEntrypoint {
+	void generateClientResourcePack(ArtificeResourcePack.ClientResourcePackBuilder builder);
+	ResourceLocation getClientID();
+}