+++ src/main/java/io/github/vampirestudios/artifice/api/ArtificeDataPackEntrypoint.java
@@ -1,0 +1,8 @@
+package io.github.vampirestudios.artifice.api;
+
+import net.minecraft.resources.ResourceLocation;
+
+public interface ArtificeDataPackEntrypoint {
+	void generateServerResourcePack(ArtificeResourcePack.ServerResourcePackBuilder builder);
+	ResourceLocation getServerID();
+}